feat(rooch-da): add async support and segment updates in indexer
Refactored `TxPositionIndexer` and `TxMetaStore` to handle asynchronous operations and added support for segment-based updates. Simplifies the indexing logic by delegating segment handling and update logic to reusable methods.
